split up large save_saga module
