+++ frontend/javascripts/test/sagas/proofreading/proofreading_multi_user.spec.ts
@@ -21,7 +21,7 @@ import {
 } from "viewer/model/actions/volumetracing_actions";
 import type { Saga } from "viewer/model/sagas/effect_generators";
 import { hasRootSagaCrashed } from "viewer/model/sagas/root_saga";
-import { VERSION_POLL_INTERVAL_COLLAB } from "viewer/model/sagas/saving/save_saga";
+import { VERSION_POLL_INTERVAL_COLLAB } from "viewer/model/sagas/saving/version_poll_saga";
 import { Store } from "viewer/singletons";
 import { startSaga } from "viewer/store";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
