+++ frontend/javascripts/viewer/store.ts
@@ -223,7 +223,7 @@ export type SegmentJournalEntry = {
 // Note that VolumeTracing should only contain state that is persisted on the
 // server (i.e., state that is synced via the save queue). This is important
 // because the VolumeTracing objects are stashed and restored from
-// RebaseRelevantAnnotationState during rebasing (see save_saga.tsx). Any
+// RebaseRelevantAnnotationState during rebasing (see rebasing_sagas.ts). Any
 // state that is not synced would be reset to the last synced version on
 // every rewinding rebase (see #9559). Local-only state belongs into
 // `state.localSegmentationStateByLayer` instead.
@@ -712,7 +712,7 @@ export type LocalSegmentationState = {
   // Note, that it is intentional that the marker position is stored here (in the
   // user-local, per-layer state) instead of within the VolumeTracing. The VolumeTracing
   // objects are stashed and restored from RebaseRelevantAnnotationState during rebasing
-  // (see save_saga.tsx). Storing the marker position there would reset it to the position
+  // (see rebasing_sagas.ts). Storing the marker position there would reset it to the position
   // of the last synced version on every rewinding rebase (see #9559).
   readonly proofreadingMarkerPosition: Vector3 | undefined;
 };
@@ -720,7 +720,7 @@ export type LocalSegmentationState = {
 // LocalAnnotationState holds local, non-persisted state that applies to the whole annotation
 // (in contrast to LocalSegmentationState, which is scoped to a single segmentation layer, and
 // in contrast to StoreAnnotation, which mirrors the persisted/synced annotation and is stashed
-// and restored during rebasing, see save_saga.tsx).
+// and restored during rebasing, see rebasing_sagas.ts).
 export type LocalAnnotationState = {
   // Bounding boxes are shared/mirrored across all tracings of an annotation (see
   // updateUserBoundingBoxes in annotation_reducer.ts), so their id reservations are
